# Unsafe vector lookup optimisation